move resource test locations
